Organize pages and components with index.js
