+++ src/pages/index.js
@@ -1,0 +1,8 @@
+export { default as AppInitStatus } from "./AppInitStatus";
+export { default as CatchDetail } from "./CatchDetail";
+export { default as CruiseDetail } from "./CruiseDetail";
+export { default as CruiseList } from "./CruiseList";
+export { default as CruiseNew } from "./CruiseNew";
+export { default as Login } from "./Login";
+export { default as StationDetail } from "./StationDetail";
+export { default as SwitchAccounts } from "./SwitchAccounts";
